Refactor/inputs

+++ src/main/java/com/todo/back/services/TodoService.java
@@ -4,6 +4,10 @@ import com.todo.back.controller.TodoItemController;
 import com.todo.back.dto.TodoDto;
 import com.todo.back.model.TodoItem;
 import com.todo.back.repository.ItemRepository;
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.ConstraintViolationException;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
 import org.springframework.hateoas.CollectionModel;
 import org.springframework.hateoas.EntityModel;
 import org.springframework.stereotype.Service;
@@ -13,6 +17,7 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.Date;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
@@ -23,8 +28,11 @@ public class TodoService {
 
     private final ItemRepository repository;
 
+    private final Validator validator;
+
     public TodoService(ItemRepository repository) {
         this.repository = repository;
+        this.validator = Validation.buildDefaultValidatorFactory().getValidator();
     }
 
     public CollectionModel<EntityModel<TodoItem>> todos() {
@@ -72,6 +80,11 @@ public class TodoService {
     public TodoItem addTodo(TodoDto todoDto) {
 
         TodoItem todo = getTodoFromDto(todoDto);
+        Set<ConstraintViolation<String>> violations = validator.validate(todo.getContent());
+
+        if (todo.getContent().length() > 100) {
+            throw new ConstraintViolationException("The content is too long", violations);
+        }
 
         todo.setDone(false);
 
@@ -79,6 +92,10 @@ public class TodoService {
     }
 
     public TodoItem editTodoStatus(TodoDto todoDto) {
+
+        if (todoDto == null) {
+            throw new IllegalArgumentException("Todo is invalid, can't edit the status");
+        }
 
         return repository.findById(todoDto.getId()) //
                 .map(todo -> {
@@ -95,6 +112,10 @@ public class TodoService {
     }
 
     public boolean deleteTodo(String id) {
+
+        if (id == null) {
+            throw new IllegalArgumentException("Id is missing, can't delete");
+        }
 
         repository.deleteById(id);
